==> src/01_gene_list_prep.Rmd <==
---
title: "1. create gene signatures from the excel files"
#output: html_notebook
---

## Initial setting
### Load libraries
```{r}
suppressPackageStartupMessages({
  library(openxlsx)
})
```

### Set paths
This part is moved to  `.Rprofile` and not not shared in github. (See README.md)

## Load manually cureated gene lists
### Funciton to load genes
```{r}
load_genes <- function(sheet_name) {
  genes <- read.xlsx(file.path(table_dir, "Table S1 gene reference.xlsx"), sheet = sheet_name)
  unique_genes <- unique(genes$Gene)
  return(unique_genes)
}
```

### Load ERS, APM, and TC genes
```{r}
ers_gns <- load_genes("ERS gene list")
apm_gns <- load_genes("APM gene list")
tc_gns <- load_genes("TC gene list")

# Output gene counts (Optional)
length(ers_gns) # 496 genes
length(apm_gns) # 324 genes
length(tc_gns) # 221 genes
```


## Save the gene lists as an rda file
```{r}
gene.sig <- list(ERS = ers_gns, APM = apm_gns, TC = tc_gns)
save(gene.sig,file=file.path(obj_dir,"1-4_prep","gene_signature.rda"))
```
