==> src/06_TCGA_METABRIC_subtypes.Rmd <==
---
title: "6. Ccorrelation in different breast cancer subtypes"
# output: html_notebook
---

## Initial setup

### Load libraries
```{r}
suppressPackageStartupMessages({
  library(RColorBrewer)
  library(ggplot2)
  library(heatmap3)
  library(GSVA)
  library(tidyr)
  library(ROCR)
  library(MASS)
  library(dplyr)
})
source(file.path(src_dir,"utils","plot_module_scores.R"))
source(file.path(src_dir,"utils","subtype_violin_plot.R"))
```


### Load data
```{r}
# Load required data
x <- load(file.path(obj_dir, "TCGA_META", "tcga_processed.rda"))
load(file.path(obj_dir, "TCGA_META", "metabric_processed.rda"))
load(file.path(obj_dir, "TCGA_META", "tcga_metabric_cors.rda"))
load(file.path(obj_dir, "misc", "heatmap_sidecols.rda"))

tested.gns <- readRDS(file.path(obj_dir, "TCGA_META", "common_tested_gns.rds"))

sigs <- readRDS(file.path(obj_dir, "Gene_modules", "modules3_ers_immune_cc.rds"))
mod.gns <- readRDS(file.path(obj_dir,"Gene_modules","module_genes_final.rds"))
comm.gns <- unlist(sigs[c(1,3)])

hallmarks <- c("ESR1", "CD8A", "HLA-A")
```

## Heatmap of Correlation for Subtypes

### TCGA

#### [Fig 1D] Heatmap of 254 module gene correlates
```{r}
## subtypes
ttls <- c("HR+/HER2-","HR-/HER2+","HR+/HER2+","HR-/HER2-")
smpl.class <- names(tcga.sp)
names(ttls) <- smpl.class

tmp <- tcga.sp$hr[comm.gns,comm.gns]

cols <- colorRampPalette(c("navy", "white", "firebrick3"))(1024)

hc <- heatmap3(tmp,
         # Colv=dend1,Rowv=dend1,
         ColSideColors=lcols[comm.gns,c(1,3,4)],
         col=cols,
         scale="none",
         balanceColor=TRUE,
         main=paste0(ttls[1],", (TCGA)"))
comm.gns <- comm.gns[hc$hcr$order]

for(cl in smpl.class){
  pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","heatmap",paste0("tcga_",cl,"_v4.pdf")),width=7,height=7)	
	par(cex.main=3)

	tmp1 <- tcga.sp[[cl]][comm.gns,comm.gns]

	cgs <- comm.gns
	cgs[!cgs %in% hallmarks] <- NA
	
	rownames(tmp1) <- colnames(tmp1) <- cgs
  heatmap3(tmp1,
           Colv=NA,Rowv=NA,
           # showColDendro=FALSE,showRowDendro=FALSE,
           ColSideColors=lcols[comm.gns,c(1,3,4)],
           col=cols,
           scale="none",
           balanceColor=TRUE,
           main=paste0(ttls[cl],", (TCGA)"))
	dev.off()
}

```

#### [Fig 1E] Heatmap of mean module gene correlates
```{r}
mean.tcga.sp <- list()
cols <- colorRampPalette(c("navy", "white", "firebrick3"))(201)

for(cl in smpl.class){
  tmp <- tcga.sp[[cl]][comm.gns,comm.gns]
  diag(tmp) <- NA
  
  mean.tcga.sp[[cl]] <- sapply(mod.gns,function(x){
    sapply(mod.gns,function(y){
      mean(tmp[x,y],na.rm=T)
    })
  })

  idx <- 1:4
  
  n.max.cor.1 <- round(max(abs(mean.tcga.sp[[cl]]))*100)
  cols1 <- cols[(-100:100) %in% seq(-n.max.cor.1,n.max.cor.1)]
  if(1){
    pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","heatmap",paste0("mean_tcga_",cl,"_v2.pdf")),
        width=7,height=7)
    hc.t <- heatmap3(mean.tcga.sp[[cl]][idx,idx],Rowv=NA,Colv=NA,
              col=cols1,scale="none",distfun=dist,
              ColSideColors=slcols[idx,2:1],
              RowSideColors=slcols[idx,],
              balanceColor = TRUE,
              margins=c(7,7),
              cexRow=1.5,cexCol=1.5,
              main=paste0("TCGA (mean correlation),",cl))
    dev.off()
  }
}
```


### METABRIC

#### [Fig S2C] Heatmap of individual genes
```{r}
tmp <- meta.sp$hr[comm.gns,comm.gns]

cols <- colorRampPalette(c("navy", "white", "firebrick3"))(1024)

for(cl in smpl.class){
  pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","heatmap",paste0("meta_",cl,"_v2.pdf")),width=7,height=7)	
	par(cex.main=3)

	tmp <- meta.sp[[cl]][comm.gns,comm.gns]

	cgs <- comm.gns
	cgs[!cgs %in% hallmarks] <- NA
	
	rownames(tmp) <- colnames(tmp) <- cgs
  heatmap3(tmp,
           Colv=NA,Rowv=NA,
           ColSideColors=lcols[comm.gns,c(1,3,4)],
           col=cols,
           scale="none",
           balanceColor=TRUE,
           main=paste0(ttls[cl]," (METABRIC)"))
  dev.off()
}

```

#### [Fig S2D] heatmap of mean correlations
```{r}
mean.meta.sp <- list()
cols <- colorRampPalette(c("navy", "white", "firebrick3"))(201)

for(cl in smpl.class){
  tmp <- meta.sp[[cl]][comm.gns,comm.gns]
  diag(tmp) <- NA
  
  mean.meta.sp[[cl]] <- sapply(mod.gns,function(x){
    sapply(mod.gns,function(y){
      mean(tmp[x,y],na.rm=T)
    })
  })

  idx <- 1:4

  n.max.cor.1 <- round(max(abs(mean.meta.sp[[cl]]))*100)
  cols1 <- cols[(-100:100) %in% seq(-n.max.cor.1,n.max.cor.1)]
  
  pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","heatmap",paste0("mean_meta_",cl,"_v2.pdf")),
      width=7,height=7)
  hc.t <- heatmap3(mean.meta.sp[[cl]][idx,idx],Rowv=NA,Colv=NA,
            col=cols1,scale="none",distfun=dist,
            ColSideColors=slcols[idx,2:1],
            RowSideColors=slcols[idx,],
            balanceColor = TRUE,
            margins=c(7,7),
            cexRow=1.5,cexCol=1.5,
            main=paste0("METABRIC (mean correlation),",cl))
  dev.off()
}
```

## Violin plot
### TCGA
#### Spearman correlation
```{r}
tcga.sp.panel <- list()
tcga.sp.panel$ers <- 
  tcga.sp.panel[["TNFa/NFkB"]] <- 
  tcga.sp.panel[["IFN-I"]] <- 
  tcga.sp.panel[["APM/TC"]] <- 
  tcga.sp.panel[["Immune"]] <- 
  tcga.sp.panel$ers.imm1 <- 
  tcga.sp.panel$ers.imm2 <- 
  tcga.sp.panel$ers.imm3 <-   
  tcga.sp.panel$ers.imm <- list()

tcga.sp.panel <- tcga.sp.panel[9:1]
tcga.pvals1 <- tcga.pvals <- list()

imm.gns <- unlist(mod.gns[-1])

n <- 406 # number of SCC within ERS

for(cl in smpl.class[c(1,3,4,2)]){
	tmp <- tcga.sp[[cl]]
	diag(tmp) <- NA

	g1 <- tmp[mod.gns[["ERS"]],mod.gns[["ERS"]]] 
	g1 <- g1[lower.tri(g1)] # 406
	g2 <- tmp[mod.gns[["TNFa/NFkB"]],mod.gns[["TNFa/NFkB"]]] 
	g2 <- g2[lower.tri(g2)] # 1378
	g3 <- tmp[mod.gns[["IFN-I"]],mod.gns[["IFN-I"]]] 
	g3 <- g3[lower.tri(g3)] # 378
	g4 <- tmp[mod.gns[["APM/TC"]],mod.gns[["APM/TC"]]] 
	g4 <- g4[lower.tri(g4)] # 10153
	g5 <- tmp[imm.gns,imm.gns]
	g5 <- g5[lower.tri(g5)] # 25200
	g6 <- as.vector(tmp[mod.gns[["ERS"]],mod.gns[["TNFa/NFkB"]]])# 1537
	g7 <- as.vector(tmp[mod.gns[["ERS"]],mod.gns[["IFN-I"]]])# 841
	g8 <- as.vector(tmp[mod.gns[["ERS"]],mod.gns[["APM/TC"]]])# 4147
	g9 <- as.vector(tmp[mod.gns[["ERS"]],imm.gns]) # 6525

	set.seed(1)	
	g1 <- sample(g1,n)
	g2 <- sample(g2,n)
	g3 <- sample(g3,n)
	g4 <- sample(g4,n)
	g5 <- sample(g5,n)
	g6 <- sample(g6,n)
	g7 <- sample(g7,n)
	g8 <- sample(g8,n)
	g9 <- sample(g9,n)	
	
	tcga.sp.panel$ers[[cl]] <- g1
	tcga.sp.panel[["TNFa/NFkB"]][[cl]] <- g2
	tcga.sp.panel[["IFN-I"]][[cl]] <- g3 
	tcga.sp.panel[["APM/TC"]][[cl]] <- g4
	tcga.sp.panel[["Immune"]][[cl]] <- g5
	tcga.sp.panel$ers.imm1[[cl]] <- g6 
	tcga.sp.panel$ers.imm2[[cl]] <- g7 
	tcga.sp.panel$ers.imm3[[cl]] <- g8 
	tcga.sp.panel$ers.imm[[cl]] <- g9 
	
	pvals <- sapply(1:9,function(i){
	  obj <- get(paste0("g",i))
	  wilcox.test(obj)$p.value
	})
	pvals1 <- sapply(1:9,function(i){
	  obj <- get(paste0("g",i))
	  t.test(obj)$p.value
	})

	names(pvals) <- names(pvals1) <- c("ERS","TNFA/NFKB","IFN-I","APM/TC","IMMUNE","ERS.IMM1","ERS.IMM2","ERS.IMM3","ERS.IMM")
	
	tcga.pvals[[cl]] <- pvals
	tcga.pvals1[[cl]] <- pvals1
}

tcga.pvals <- do.call(rbind,tcga.pvals) # wilcox
tcga.pvals1 <- do.call(rbind,tcga.pvals1) # t.test

subtypes <- c("HR+/HER2-","HR+/HER2+","HR-/HER2-","HR-/HER2+") 

names(tcga.sp.panel$ers) <- 
  names(tcga.sp.panel[["TNFa/NFkB"]]) <- 
  names(tcga.sp.panel[["IFN-I"]]) <- 
  names(tcga.sp.panel[["APM/TC"]]) <- 
  names(tcga.sp.panel[["Immune"]]) <-   
  names(tcga.sp.panel$ers.imm1) <- 
  names(tcga.sp.panel$ers.imm2) <- 
  names(tcga.sp.panel$ers.imm3) <- 
  names(tcga.sp.panel$ers.imm) <- subtypes

tcga.sps <- lapply(seq(tcga.sp.panel),function(i){
  n <- names(tcga.sp.panel)[[i]]
  lst <- tcga.sp.panel[[n]]
  tmp <- data.frame(do.call(cbind,lst))
  names(tmp) <- 1:4
  tmp <- tmp %>%
    tidyr::gather("subtype","sp",1:4) %>%
    mutate(type=n)
  return(tmp)
})

tcga.sps <- do.call(rbind,tcga.sps) 
tcga.sps <- tcga.sps %>%
  mutate(subtype=subtypes[as.numeric(subtype)]) %>%
  mutate(subtype=factor(subtype,levels=subtypes)) %>%
  mutate(type=toupper(type))

tcga.ps <- array(rowSums(sapply(c(25,50,100,200),function(i)-log10(tcga.pvals1)>i)),dim(tcga.pvals1)) # t-test
colnames(tcga.ps) <- toupper(names(tcga.sp.panel))
rownames(tcga.ps) <- rownames(tcga.pvals1)

df.txt <- data.frame(tcga.ps) %>%
  tibble::rownames_to_column("subtype") %>%
  mutate(subtype=factor(subtype,levels=c("hr","dp","tnbc","her2"))) %>%
  tidyr::gather("type","txt",-1) %>%
  mutate(txt=sapply(txt,function(i)paste(rep("*",i),collapse=""))) %>%
  mutate(is.hr=factor(subtype %in% c("hr","dp"),levels=c("TRUE","FALSE")))  %>%
  mutate()
    
levels(df.txt$subtype) <- subtypes

dfx.tcga <- tcga.sps %>% 
  mutate(type = sub("[-/]",".",type)) %>%
  left_join(df.txt,by=c("subtype","type"))
dfx1 <- dfx.tcga %>% select(-2) %>% unique()
```


#### [Fig S1G-I] Violin plot

```{r}
p <- subtype_violin_within(dfx=dfx.tcga,mod="ERS") + ggtitle("ERS (TCGA)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_ers_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_within(dfx=dfx.tcga,mod="IMMUNE") + ggtitle("Pan-immune (TCGA)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_imms_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_within(dfx=dfx.tcga,mod="TNFA.NFKB") + ggtitle("TNFa/NFkB (TCGA)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_imm1_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_within(dfx=dfx.tcga,mod="IFN.I") + ggtitle("IFN-I (TCGA)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_imm2_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_within(dfx=dfx.tcga,mod="APM.TC") + ggtitle("APM/TC (TCGA)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_imm3_v2.pdf"),width=3,height=6)
print(p)
dev.off()
```


```{r}
p <- subtype_violin_between(dfx=dfx.tcga,mod="IMMUNE") + ggtitle("ERS vs Pan-immune (TCGA)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_ers_imm_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_between(dfx=dfx.tcga,mod="TNFA.NFKB") + ggtitle("ERS vs TNFa/NFkB (TCGA)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_ers_imm1_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_between(dfx=dfx.tcga,mod="IFN.I") + ggtitle("ERS vs IFN-I (TCGA)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_ers_imm2_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_between(dfx=dfx.tcga,mod="APM.TC") + ggtitle("ERS vs APM/TC (TCGA)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_ers_imm3_v2.pdf"),width=3,height=6)
print(p)
dev.off()
```

### METABRIC
#### Spearman correlation
```{r}
meta.sp.panel <- list()
meta.sp.panel$ers <- 
  meta.sp.panel[["TNFa/NFkB"]] <- 
  meta.sp.panel[["IFN-I"]] <- 
  meta.sp.panel[["APM/TC"]] <- 
  meta.sp.panel[["Immune"]] <- 
  meta.sp.panel$ers.imm1 <- 
  meta.sp.panel$ers.imm2 <- 
  meta.sp.panel$ers.imm3 <-   
  meta.sp.panel$ers.imm <- list()

meta.sp.panel <- meta.sp.panel[9:1]
meta.pvals1 <- meta.pvals <- list()

imm.gns <- unlist(mod.gns[-1])

n <- 406 # number of SCC within ERS

for(cl in smpl.class[c(1,3,4,2)]){
	tmp <- meta.sp[[cl]]
	diag(tmp) <- NA

	g1 <- tmp[mod.gns[["ERS"]],mod.gns[["ERS"]]] 
	g1 <- g1[lower.tri(g1)] # 406
	g2 <- tmp[mod.gns[["TNFa/NFkB"]],mod.gns[["TNFa/NFkB"]]] 
	g2 <- g2[lower.tri(g2)] # 1378
	g3 <- tmp[mod.gns[["IFN-I"]],mod.gns[["IFN-I"]]] 
	g3 <- g3[lower.tri(g3)] # 378
	g4 <- tmp[mod.gns[["APM/TC"]],mod.gns[["APM/TC"]]] 
	g4 <- g4[lower.tri(g4)] # 10153
	g5 <- tmp[imm.gns,imm.gns]
	g5 <- g5[lower.tri(g5)] # 25200
	g6 <- as.vector(tmp[mod.gns[["ERS"]],mod.gns[["TNFa/NFkB"]]])# 1537
	g7 <- as.vector(tmp[mod.gns[["ERS"]],mod.gns[["IFN-I"]]])# 841
	g8 <- as.vector(tmp[mod.gns[["ERS"]],mod.gns[["APM/TC"]]])# 4147
	g9 <- as.vector(tmp[mod.gns[["ERS"]],imm.gns]) # 6525

	set.seed(1)	
	g1 <- sample(g1,n)
	g2 <- sample(g2,n)
	g3 <- sample(g3,n)
	g4 <- sample(g4,n)
	g5 <- sample(g5,n)
	g6 <- sample(g6,n)
	g7 <- sample(g7,n)
	g8 <- sample(g8,n)
	g9 <- sample(g9,n)	
	
	meta.sp.panel$ers[[cl]] <- g1
	meta.sp.panel[["TNFa/NFkB"]][[cl]] <- g2
	meta.sp.panel[["IFN-I"]][[cl]] <- g3 
	meta.sp.panel[["APM/TC"]][[cl]] <- g4
	meta.sp.panel[["Immune"]][[cl]] <- g5
	meta.sp.panel$ers.imm1[[cl]] <- g6 
	meta.sp.panel$ers.imm2[[cl]] <- g7 
	meta.sp.panel$ers.imm3[[cl]] <- g8 
	meta.sp.panel$ers.imm[[cl]] <- g9 
	
	pvals <- sapply(1:9,function(i){
	  obj <- get(paste0("g",i))
	  wilcox.test(obj)$p.value
	})
	pvals1 <- sapply(1:9,function(i){
	  obj <- get(paste0("g",i))
	  t.test(obj)$p.value
	})

	names(pvals) <- names(pvals1) <- c("ERS","TNFA/NFKB","IFN-I","APM/TC","IMMUNE","ERS.IMM1","ERS.IMM2","ERS.IMM3","ERS.IMM")
	
	meta.pvals[[cl]] <- pvals
	meta.pvals1[[cl]] <- pvals1
}

meta.pvals <- do.call(rbind,meta.pvals) # wilcox
meta.pvals1 <- do.call(rbind,meta.pvals1) # t.test

subtypes <- c("HR+/HER2-","HR+/HER2+","HR-/HER2-","HR-/HER2+") 

names(meta.sp.panel$ers) <- 
  names(meta.sp.panel[["TNFa/NFkB"]]) <- 
  names(meta.sp.panel[["IFN-I"]]) <- 
  names(meta.sp.panel[["APM/TC"]]) <- 
  names(meta.sp.panel[["Immune"]]) <-   
  names(meta.sp.panel$ers.imm1) <- 
  names(meta.sp.panel$ers.imm2) <- 
  names(meta.sp.panel$ers.imm3) <- 
  names(meta.sp.panel$ers.imm) <- subtypes

meta.sps <- lapply(seq(meta.sp.panel),function(i){
  n <- names(meta.sp.panel)[[i]]
  lst <- meta.sp.panel[[n]]
  tmp <- data.frame(do.call(cbind,lst))
  names(tmp) <- 1:4
  tmp <- tmp %>%
    tidyr::gather("subtype","sp",1:4) %>%
    mutate(type=n)
  return(tmp)
})

meta.sps <- do.call(rbind,meta.sps) 
meta.sps <- meta.sps %>%
  mutate(subtype=subtypes[as.numeric(subtype)]) %>%
  mutate(subtype=factor(subtype,levels=subtypes)) %>%
  mutate(type=toupper(type))

meta.ps <- array(rowSums(sapply(c(25,50,100,200),function(i)-log10(meta.pvals1)>i)),dim(meta.pvals1)) # t-test
colnames(meta.ps) <- toupper(names(meta.sp.panel))
rownames(meta.ps) <- rownames(meta.pvals1)

df.txt <- data.frame(meta.ps) %>%
  tibble::rownames_to_column("subtype") %>%
  mutate(subtype=factor(subtype,levels=c("hr","dp","tnbc","her2"))) %>%
  tidyr::gather("type","txt",-1) %>%
  mutate(txt=sapply(txt,function(i)paste(rep("*",i),collapse=""))) %>%
  mutate(is.hr=factor(subtype %in% c("hr","dp"),levels=c("TRUE","FALSE")))  %>%
  mutate()
    
levels(df.txt$subtype) <- subtypes

dfx.meta <- meta.sps %>% 
  mutate(type = sub("[-/]",".",type)) %>%
  left_join(df.txt,by=c("subtype","type"))
dfx1 <- dfx.meta %>% select(-2) %>% unique()
```


#### [Fig S1G-I] Violin plot
```{r}
p <- subtype_violin_within(dfx=dfx.meta,mod="ERS") + ggtitle("ERS (METABRIC)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_ers_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_within(dfx=dfx.meta,mod="IMMUNE") + ggtitle("Pan-immune (METABRIC)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_imms_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_within(dfx=dfx.meta,mod="TNFA.NFKB") + ggtitle("TNFa/NFkB (METABRIC)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_imm1_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_within(dfx=dfx.meta,mod="IFN.I") + ggtitle("IFN-I (METABRIC)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_imm2_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_within(dfx=dfx.meta,mod="APM.TC") + ggtitle("APM/TC (METABRIC)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_imm3_v2.pdf"),width=3,height=6)
print(p)
dev.off()
```


```{r}
p <- subtype_violin_between(dfx=dfx.meta,mod="IMMUNE") + ggtitle("ERS vs Pan-immune (METABRIC)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_ers_imm_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_between(dfx=dfx.meta,mod="TNFA.NFKB") + ggtitle("ERS vs TNFa/NFkB (METABRIC)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_ers_imm1_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_between(dfx=dfx.meta,mod="IFN.I") + ggtitle("ERS vs IFN-I (METABRIC)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_ers_imm2_v2.pdf"),width=3,height=6)
print(p)
dev.off()

p <- subtype_violin_between(dfx=dfx.meta,mod="APM.TC") + ggtitle("ERS vs APM/TC (METABRIC)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_ers_imm3_v2.pdf"),width=3,height=6)
print(p)
dev.off()
```

###  ESR1 expression per subclass

#### 1D/2D plots - functions


### [Fig S2E] METABRIC
```{r}
ers.exp <- meta.m3[match("ESR1",meta.syms),unlist(meta.ids[c(1,3,4,2)])]
cls <- factor(rep(subtypes,sapply(meta.ids[c(1,3,4,2)],length)),
	levels=subtypes)

df.m <- data.frame(ESR1=ers.exp,mod=cls) %>%
  mutate(is.hr = factor(mod %in% c("HR+/HER2-","HR+/HER2+"),levels=c(TRUE,FALSE)))

exp.pos <- (df.m %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))$ESR1
exp.neg <- (df.m %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))$ESR1

predictions <- c(exp.pos,exp.neg)
labels <- rep(c(1,-1),c(length(exp.pos),length(exp.neg)))

th <- threshold1(predictions,labels)

p <- my_plot_d1(data=df.m,x1="ESR1") + ggtitle("ESR1 gene (METABRIC)") + ylab("Expression (log, z-score)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_esr1_bc_subtypes.pdf"),width=4,height=7)
print(p)
dev.off()

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_esr1_bc_subtypes.pdf"),width=4,height=7)
p <- ggplot(df.m,aes(class,esr1)) + 
	geom_violin(fill='grey80') +
	geom_jitter(height = 0, width = 0.2,size=.5) +
	labs(title="ESR1 gene (METABRIC)",
        x ="", y = "Expression (log, z-score)") +
  geom_hline(yintercept = th,linetype="dashed") +
  theme_bw() +
  theme(plot.title=element_text(hjust=0.5),
        text = element_text(size = 15),
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))    
p
dev.off()
```

## GSVA

### Run gsva for both TCGA and METABRIC
```{r}
gs <- lapply(c(sigs,mod.gns[-1]),identity)

mth <- "zscore"

lst.tcga.gs <- lst.meta.gs <- list()

lst.tcga.gs[[mth]] <-  GSVA::gsva(expr = tcga.m3[tested.gns,],
             gset.idx.list = gs, method = mth,
             min.sz = 10, max.sz = 500, kcdf = "Gaussian")

lst.meta.gs[[mth]] <-  GSVA::gsva(expr = meta.m3[tested.gns,],
             gset.idx.list = gs, method = mth,
             min.sz = 10, max.sz = 500, kcdf = "Gaussian")

if(0){
  # save(lst.tcga.gs,lst.meta.gs,file=file.path(obj_dir,"TCGA_META","tcga_meta_ssgsea_scores_v2.rda"))
  save(lst.tcga.gs,lst.meta.gs,file=file.path(obj_dir,"TCGA_META","tcga_meta_ssgsea_scores_zscore.rda"))  
}else{
  x <- load(file.path(obj_dir,"TCGA_META","tcga_meta_ssgsea_scores_zscore.rda"))  
}
```

### [Fig S1F] ERS vs CC
```{r}
df.gs <- t(lst.tcga.gs[["zscore"]])  %>% 
  as.data.frame() %>%
  tibble::rownames_to_column("smpl")

df.t <- data.frame(ESR1=tcga.m3["ESR1",unlist(tcga.ids[c(1,3,4,2)])],
                   class=factor(rep(names(tcga.ids[c(1,3,4,2)]),sapply(tcga.ids[c(1,3,4,2)],length)),
                                levels=c("hr","dp","tnbc","her2"),
                                labels=subtypes))# %>%


df.t1 <- df.t %>% 
  tibble::rownames_to_column("smpl") %>% 
  left_join(df.gs,by="smpl") %>%
  filter(class == "HR+/HER2-")

scc <- round(cor(df.t1$ERS,df.t1$CC,method="spearman"),2)

dc <- densCols(df.t1[,c("ERS","CC")])

##
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","tcga_ers_vs_cc_hr.pdf"),width=7,height=7)
par(mar=c(5,5,3,3))
plot(df.t1[,"ERS"], df.t1[,"CC"],pch=20,col=dc,
     main=paste0("ERS vs CC (HR+/HER2-,TCGA)"),
     xlab="ERS (GSVA z-score)",ylab="CC (GSVA z-score)")
text(-7,-18,paste0("SCC=",scc))
dev.off()
```

### TCGA - GSVA plot

#### ESR1 threshold
```{r}
esr1.exp <- tcga.m3[match("ESR1",tcga.syms),unlist(tcga.ids[c(1,3,4,2)])]

df0 <- data.frame(ESR1=esr1.exp) %>% tibble::rownames_to_column("id")
df2 <- data.frame(id=unlist(tcga.ids),mod=rep(names(tcga.ids),sapply(tcga.ids,length))) %>%
  left_join(df0,by="id") %>%
  mutate(mod=factor(mod,levels=c("hr","dp","tnbc","her2")))
levels(df2$mod) <- c("HR+/HER2-","HR+/HER2+","HR-/HER2-","HR-/HER2+")

exp.pos <- (df2 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))$ESR1
exp.neg <- (df2 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))$ESR1

predictions <- c(exp.pos,exp.neg)
labels <- rep(c(1,-1),c(length(exp.pos),length(exp.neg)))
pred <- prediction(predictions,labels)
perf <- performance(pred, "tpr", "fpr")

auc_perf <- ROCR::performance(pred,"auc")
auroc <- auc_perf@y.values[[1]]

th <- threshold1(predictions,labels) # threshold for ESR1
``` 

```{r}
tcga.gs <- lst.tcga.gs[["zscore"]] # gsva or zscore

df0 <- data.frame(t(tcga.gs)) %>% tibble::rownames_to_column("id")
df1 <- data.frame(id=unlist(tcga.ids),mod=rep(names(tcga.ids),sapply(tcga.ids,length))) %>%
  left_join(df0,by="id") %>%
  mutate(mod=factor(mod,levels=c("hr","dp","tnbc","her2"))) %>%
  mutate(is.hr=factor(mod %in% c("hr","dp"),levels=c("TRUE","FALSE"))) %>%
  left_join(df2[c(1,3)],by="id")
levels(df1$mod) <- c("HR+/HER2-","HR+/HER2+","HR-/HER2-","HR-/HER2+")

exp.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))$ERS
exp.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))$ERS

predictions <- c(exp.pos,exp.neg)
labels <- rep(c(1,-1),c(length(exp.pos),length(exp.neg)))
pred <- prediction(predictions,labels)

perf <- performance(pred, "tpr", "fpr")

auc_perf <- ROCR::performance(pred,"auc")
auroc <- auc_perf@y.values[[1]]

th.ers <- threshold1(predictions,labels) # threshold for ERS
```


```{r}
x.th <- "ESR1"
exp.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))[[x.th]]
exp.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))[[x.th]]

predictions <- c(exp.pos,exp.neg)
labels <- rep(c(1,-1),c(length(exp.pos),length(exp.neg)))
if(x.th %in% c("ERS","ESR1")){
  labels <- labels
}else{
  labels <- -labels
}
th.esr1 <- threshold1(predictions,labels)

df1 <- df1 %>%
  mutate(ers.pos = ERS > th.ers) %>%
  mutate(esr.pos = ESR1 > th.esr1)
```

#### [Fig. 1C] ESR1 gene exprssion

```{r}
p <- my_plot_d1(data=df1,x1="ESR1") + ggtitle("ESR1 gene (TCGA)") + ylab("Expression (log, z-score)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_violin_esr1.pdf"),width=4,height=8)
print(p)
dev.off()
```
#### [Fig. 1D] ESR1 gene vs ERS module
```{r}
lm1 <- lm(ERS ~ ESR1,data=df1 %>% filter(!ers.pos))
lm2 <- lm(ERS ~ ESR1,data=df1 %>% filter(ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=-2,y=2,label=paste0("ERS-low:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=0.7,y=-6,label=paste0("ERS-high:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3_th(data=df1,x1="ESR1",x2="ERS",draw.th=T,x.th="ESR1") +
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","tcga_esr1_vs_ers_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```

#### [Fig. 1F] between immune modules
```{r}
ers.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))$ERS
ers.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))$ERS

lm1 <- lm(APM.TC ~ TNFa.NFkB, data=df1 %>% filter(ers.pos))
lm2 <- lm(APM.TC ~ TNFa.NFkB, data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=-12,y=10,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=7,y=-15,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="TNFa.NFkB",x2="APM.TC",draw.th=F,x.th="ERS",split.lm=TRUE)+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","tcga_imm1_vs_imm3_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```


```{r}
lm1 <- lm(APM.TC ~ IFN.I,data=df1 %>% filter(ers.pos))
lm2 <- lm(APM.TC ~ IFN.I,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=-10,y=10,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=7,y=-18,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="IFN.I",x2="APM.TC",draw.th=F,x.th="ERS")+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","tcga_imm2_vs_imm3_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```

#### [Fig. 1G] between immune modules

```{r}
lm1 <- lm(IFN.I ~ TNFa.NFkB,data=df1 %>% filter(ers.pos))
lm2 <- lm(IFN.I ~ TNFa.NFkB,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=-12,y=10,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=7,y=-10,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3_th(data=df1,x1="TNFa.NFkB",x2="IFN.I",draw.th=T,x.th="ERS") +
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","tcga_imm1_vs_imm2_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```

```{r}
th <- function(x){
  exp.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))[[x]]
  exp.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))[[x]]
  
  predictions <- c(exp.pos,exp.neg)
  labels <- rep(c(1,-1),c(length(exp.pos),length(exp.neg)))
  
  if(x %in% c("ERS","ESR1")){
    labels <- labels
  }else{
    labels <- -labels
  }
  
  pred <- prediction(predictions,labels)
  
  perf <- performance(pred, "tpr", "fpr")
  
  th <- threshold1(predictions,labels)
  return(th)
}

th1 <- th("TNFa.NFkB")
th2 <- th("IFN.I")
df1 <- df1 %>%
  mutate(cat = c("Neg","NFkB","IFN-I","N+I")[(TNFa.NFkB > th1) + (IFN.I > th2)*2 + 1]) %>%
  mutate(cat =factor(cat,levels=c("Neg","NFkB","IFN-I","N+I")))

ucols <- slcols[,"immune"]
ucols[1] <- "grey80"
ucols[4] <- brewer.pal(4,"Accent")[4]
names(ucols) <- levels(df1$cat)

idx <- list(c(1,2),c(2,3),c(3,4),c(1,3),c(2,4),c(1,4))
levs <- levels(df1$cat)
pvals <- rep(NA,length(idx))
for(i in seq(idx)){
  this <- levels(df1$cat)[idx[[i]]]
  t1 <- df1 %>% filter(cat ==this[1]) %>% pull(APM.TC)
  t2 <- df1 %>% filter(cat ==this[2]) %>% pull(APM.TC)  
  pvals[i] <- signif(t.test(t1,t2,alternative="less")$p.value,3)
}
mid.i <- sapply(idx,mean)
ps <- data.frame(i=seq(mid.i),
                 x1=sapply(idx,function(x)x[1]) + .05,
                 x2=sapply(idx,function(x)x[2]) - .05,
                 x.mid = mid.i,
                 y1=c(1,1,1,2,3,4) * 3 + 25,
                 y2=c(1,1,1,2,3,4) * 3 + 25,
                 p.val=pvals)

p <- ggplot(df1,aes(x=cat,y=APM.TC)) +
  geom_violin(position="dodge",aes(fill=cat)) +
  scale_fill_manual(values=ucols) +
  geom_jitter(width=.2,size=.5) +
  geom_segment(data=ps,aes(x=x1,xend=x2,y=y1,yend=y2,group=i))+
  geom_text(data=ps,aes(x=x.mid,y=y1+1.2,label=p.val)) +
  theme_bw() +
  xlab("") +
  ylab("") + 
  ggtitle("") +
  theme(plot.title=element_text(hjust=0.5),
    text = element_text(size = 20),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
    legend.position="none")  

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","tcga_violin_imm.pdf"),width=5,height=7)
print(p)
dev.off()

```

#### [Fig. 1D] between ERS and ESR1

```{r}
exp.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))$ESR1
exp.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))$ESR1
  
lm1 <- lm(ERS ~ ESR1,data=df1 %>% filter(esr.pos))
lm2 <- lm(ERS ~ ESR1,data=df1 %>% filter(!esr.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=.8,y=-6,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),esr.pos=TRUE)
txt2 <- data.frame(x=-2,y=2,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),esr.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="ESR1",x2="ERS",draw.th=TRUE,x.th="ESR1",split.lm=TRUE) +
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","tcga_esr1_vs_ers_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```
#### [Fig. 1E] between ERS and immune modules

```{r}
lm1 <- lm(TNFa.NFkB ~ ERS,data=df1 %>% filter(ers.pos))
lm2 <- lm(TNFa.NFkB ~ ERS,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=5,y=15,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=-10,y=-10,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="ERS",x2="TNFa.NFkB",draw.th=TRUE,x.th="ERS",split.lm=TRUE)+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","tcga_ers_vs_imm1_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```

```{r}
lm1 <- lm(IFN.I ~ ERS,data=df1 %>% filter(ers.pos))
lm2 <- lm(IFN.I ~ ERS,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=5,y=15,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=-10,y=-10,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="ERS",x2="IFN.I",draw.th=TRUE,x.th="ERS",split.lm=TRUE)+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","tcga_ers_vs_imm2_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```

```{r}
lm1 <- lm(APM.TC ~ ERS,data=df1 %>% filter(ers.pos))
lm2 <- lm(APM.TC ~ ERS,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=5,y=15,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=-10,y=-10,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="ERS",x2="APM.TC",draw.th=TRUE,x.th="ERS",split.lm=TRUE)+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","tcga_ers_vs_imm3_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```
#### multivariate regression
```{r}
lm1 <- lm(APM.TC ~ TNFa.NFkB + IFN.I + TNFa.NFkB:IFN.I,data=df1)
summary(lm1)$coefficient
```

#### network analysis - TBD

### METABRIC - GSVA plot

#### ESR1 threshold
```{r}
esr1.exp <- meta.m3[match("ESR1",meta.syms),unlist(meta.ids[c(1,3,4,2)])]

df0 <- data.frame(ESR1=esr1.exp) %>% tibble::rownames_to_column("id")
df2 <- data.frame(id=unlist(meta.ids),mod=rep(names(meta.ids),sapply(meta.ids,length))) %>%
  left_join(df0,by="id") %>%
  mutate(mod=factor(mod,levels=c("hr","dp","tnbc","her2")))
levels(df2$mod) <- c("HR+/HER2-","HR+/HER2+","HR-/HER2-","HR-/HER2+")

exp.pos <- (df2 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))$ESR1
exp.neg <- (df2 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))$ESR1

predictions <- c(exp.pos,exp.neg)
labels <- rep(c(1,-1),c(length(exp.pos),length(exp.neg)))
pred <- prediction(predictions,labels)
perf <- performance(pred, "tpr", "fpr")

th <- threshold1(predictions,labels) # threshold for ESR1
``` 

```{r}
meta.gs <- lst.meta.gs[["zscore"]] # gsva or zscore

df0 <- data.frame(t(meta.gs)) %>% tibble::rownames_to_column("id")
df1 <- data.frame(id=unlist(meta.ids),mod=rep(names(meta.ids),sapply(meta.ids,length))) %>%
  left_join(df0,by="id") %>%
  mutate(mod=factor(mod,levels=c("hr","dp","tnbc","her2"))) %>%
  mutate(is.hr=factor(mod %in% c("hr","dp"),levels=c("TRUE","FALSE"))) %>%
  left_join(df2[c(1,3)],by="id")
levels(df1$mod) <- c("HR+/HER2-","HR+/HER2+","HR-/HER2-","HR-/HER2+")

exp.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))$ERS
exp.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))$ERS

predictions <- c(exp.pos,exp.neg)
labels <- rep(c(1,-1),c(length(exp.pos),length(exp.neg)))
pred <- prediction(predictions,labels)

perf <- performance(pred, "tpr", "fpr")

auc_perf <- ROCR::performance(pred,"auc")
auroc <- auc_perf@y.values[[1]]

th.ers <- threshold1(predictions,labels) # threshold for ERS
```


```{r}
x.th <- "ESR1"
exp.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))[[x.th]]
exp.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))[[x.th]]

predictions <- c(exp.pos,exp.neg)
labels <- rep(c(1,-1),c(length(exp.pos),length(exp.neg)))
if(x.th %in% c("ERS","ESR1")){
  labels <- labels
}else{
  labels <- -labels
}
th.esr1 <- threshold1(predictions,labels)

df1 <- df1 %>%
  mutate(ers.pos = ERS > th.ers) %>%
  mutate(esr.pos = ESR1 > th.esr1)
```

#### [Fig. S2E] ESR1 gene exprssion

```{r}
p <- my_plot_d1(data=df1,x1="ESR1") + ggtitle("ESR1 gene (METABRIC)") + ylab("Expression (log, z-score)")
pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","meta_violin_esr1.pdf"),width=4,height=8)
print(p)
dev.off()
```
#### [Fig. S2F] ESR1 gene vs ERS module
```{r}
lm1 <- lm(ERS ~ ESR1,data=df1 %>% filter(!ers.pos))
lm2 <- lm(ERS ~ ESR1,data=df1 %>% filter(ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=-2,y=2,label=paste0("ERS-low:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=0.7,y=-6,label=paste0("ERS-high:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3_th(data=df1,x1="ESR1",x2="ERS",draw.th=T,x.th="ESR1") +
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","meta_esr1_vs_ers_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```
#### [Fig. S2K] between immune modules
```{r}
ers.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))$ERS
ers.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))$ERS

lm1 <- lm(APM.TC ~ TNFa.NFkB, data=df1 %>% filter(ers.pos))
lm2 <- lm(APM.TC ~ TNFa.NFkB, data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=-10,y=10,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=7,y=-15,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="TNFa.NFkB",x2="APM.TC",draw.th=F,x.th="ERS",split.lm=TRUE)+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","meta_imm1_vs_imm3_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```


```{r}
lm1 <- lm(APM.TC ~ IFN.I,data=df1 %>% filter(ers.pos))
lm2 <- lm(APM.TC ~ IFN.I,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=-7,y=10,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=7,y=-18,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="IFN.I",x2="APM.TC",draw.th=F,x.th="ERS")+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","meta_imm2_vs_imm3_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```

#### [Fig. S2L] between immune modules

```{r}
lm1 <- lm(IFN.I ~ TNFa.NFkB,data=df1 %>% filter(ers.pos))
lm2 <- lm(IFN.I ~ TNFa.NFkB,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=-10,y=10,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=9,y=-8,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3_th(data=df1,x1="TNFa.NFkB",x2="IFN.I",draw.th=T,x.th="ERS") +
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","meta_imm1_vs_imm2_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```

```{r}
th <- function(x){
  exp.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))[[x]]
  exp.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))[[x]]
  
  predictions <- c(exp.pos,exp.neg)
  labels <- rep(c(1,-1),c(length(exp.pos),length(exp.neg)))
  
  if(x %in% c("ERS","ESR1")){
    labels <- labels
  }else{
    labels <- -labels
  }
  
  pred <- prediction(predictions,labels)
  
  perf <- performance(pred, "tpr", "fpr")
  
  th <- threshold1(predictions,labels)
  return(th)
}

th1 <- th("TNFa.NFkB")
th2 <- th("IFN.I")
df1 <- df1 %>%
  mutate(cat = c("Neg","NFkB","IFN-I","N+I")[(TNFa.NFkB > th1) + (IFN.I > th2)*2 + 1]) %>%
  mutate(cat =factor(cat,levels=c("Neg","NFkB","IFN-I","N+I")))

ucols <- slcols[,"immune"]
ucols[1] <- "grey80"
ucols[4] <- brewer.pal(4,"Accent")[4]
names(ucols) <- levels(df1$cat)

idx <- list(c(1,2),c(2,3),c(3,4),c(1,3),c(2,4),c(1,4))
levs <- levels(df1$cat)
pvals <- rep(NA,length(idx))
for(i in seq(idx)){
  this <- levels(df1$cat)[idx[[i]]]
  t1 <- df1 %>% filter(cat ==this[1]) %>% pull(APM.TC)
  t2 <- df1 %>% filter(cat ==this[2]) %>% pull(APM.TC)  
  pvals[i] <- signif(t.test(t1,t2,alternative="less")$p.value,3)
}
mid.i <- sapply(idx,mean)
ps <- data.frame(i=seq(mid.i),
                 x1=sapply(idx,function(x)x[1]) + .05,
                 x2=sapply(idx,function(x)x[2]) - .05,
                 x.mid = mid.i,
                 y1=c(1,1,1,2,3,4) * 3 + 25,
                 y2=c(1,1,1,2,3,4) * 3 + 25,
                 p.val=pvals)

p <- ggplot(df1,aes(x=cat,y=APM.TC)) +
  geom_violin(position="dodge",aes(fill=cat)) +
  scale_fill_manual(values=ucols) +
  geom_jitter(width=.2,size=.5) +
  geom_segment(data=ps,aes(x=x1,xend=x2,y=y1,yend=y2,group=i))+
  geom_text(data=ps,aes(x=x.mid,y=y1+1.2,label=p.val)) +
  theme_bw() +
  xlab("") +
  ylab("") + 
  ggtitle("") +
  theme(plot.title=element_text(hjust=0.5),
    text = element_text(size = 20),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
    legend.position="none")  

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","violin","meta_violin_imm.pdf"),width=5,height=7)
print(p)
dev.off()

```

#### [Fig. S2F] between ERS and ESR1

```{r}
exp.pos <- (df1 %>% filter(mod %in% c("HR+/HER2-","HR+/HER2+")))$ESR1
exp.neg <- (df1 %>% filter(mod %in% c("HR-/HER2-","HR-/HER2+")))$ESR1
  
lm1 <- lm(ERS ~ ESR1,data=df1 %>% filter(esr.pos))
lm2 <- lm(ERS ~ ESR1,data=df1 %>% filter(!esr.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=.8,y=-6,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),esr.pos=TRUE)
txt2 <- data.frame(x=-1.5,y=2,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),esr.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="ESR1",x2="ERS",draw.th=TRUE,x.th="ESR1",split.lm=TRUE) +
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","meta_esr1_vs_ers_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```
#### [Fig. S2G] between ERS and immune modules

```{r}
lm1 <- lm(TNFa.NFkB ~ ERS,data=df1 %>% filter(ers.pos))
lm2 <- lm(TNFa.NFkB ~ ERS,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=5,y=10,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=-7,y=-10,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="ERS",x2="TNFa.NFkB",draw.th=TRUE,x.th="ERS",split.lm=TRUE)+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","meta_ers_vs_imm1_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```

```{r}
lm1 <- lm(IFN.I ~ ERS,data=df1 %>% filter(ers.pos))
lm2 <- lm(IFN.I ~ ERS,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=5,y=12,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=-7,y=-8,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="ERS",x2="IFN.I",draw.th=TRUE,x.th="ERS",split.lm=TRUE)+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","meta_ers_vs_imm2_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```

```{r}
lm1 <- lm(APM.TC ~ ERS,data=df1 %>% filter(ers.pos))
lm2 <- lm(APM.TC ~ ERS,data=df1 %>% filter(!ers.pos))
p.val1 <- paste0("p=",signif(summary(lm1)$coefficients[2,4],3))
r.sq1 <- paste0("Adj.R^2=",signif(summary(lm1)$adj.r.squared,3))
p.val2 <- paste0("p=",signif(summary(lm2)$coefficients[2,4],3))
r.sq2 <- paste0("Adj.R^2=",signif(summary(lm2)$adj.r.squared,3))

txt1 <- data.frame(x=5,y=18,label=paste0("ERS-high:\n",p.val1,"\n",r.sq1),ers.pos=TRUE)
txt2 <- data.frame(x=-7,y=-12,label=paste0("ERS-low:\n",p.val2,"\n",r.sq2),ers.pos=FALSE)
txt <- rbind(txt1,txt2)

p <- my_plot_d3(data=df1,x1="ERS",x2="APM.TC",draw.th=TRUE,x.th="ERS",split.lm=TRUE)+
  geom_text(data=txt,aes(x=x,y=y,label=label),size=5,color=1)

pdf(file.path(fig_dir,"06_TCGA_METABRIC_subtypes","activity_plot","meta_ers_vs_imm3_v3.pdf"),width=7,height=7)
print(p)
dev.off()
```
