==> src/15_DFBCC_snSeq_prep.Rmd <==
---
title: "prep for single-nucleus RNA-seq analysis of DFBCC samples"
---

## Initial setup
### Load libraries
```{r}
suppressPackageStartupMessages({
  library(Seurat)
  library(tidyverse)
  library(RColorBrewer)
})
```

### Load input data
```{r}
obj_dir1 <- file.path(obj_dir,"DFBCC_sn")
fig_dir1 <- file.path(fig_dir,"15_DFBCC_snSeq_prep")

seurat_all <- readRDS(file=file.path(obj_dir1,"seurat_all.rds")) 
```

### Load Biomarkers
```{r}
all_gns <- rownames(GetAssayData(seurat_all))

epi_gns <- c("KRT7", "KRT8", "KRT18", "KRT19", "KRT5", "KRT6B", "KRT14", "KRT17")
fib_gns <- c("COL1A1", "ACTA2", "PDPN")
peri_gns <- c("RGS5")
endo_gns <- c("PECAM1")
adipo_gns <- c("CIDEA")
mast_gns <- c("MS4A2", "KIT")
dc_gns <- c("FCER1A", "CD1A", "IRF4", "IRF8", "IL3RA", "LAMP3")
mac_gns <- c("CD68", "CD163", "CD14", "LYZ", "FCGR3A")
t_gns <- c("CD3D", "CD2", "TRAC", "TRBC1", "TRDC", "CD4", "FOXP3", "CD8A", "GZMB", "GZMK")
nk_gns <- c("KLRD1", "KLRC1")
b_gns <- c("MS4A1", "IGHM", "JCHAIN")
pl_gns <- c("IGHA1", "IGHA2", "IGHG1", "IGHG3", "IGHG4")

lst.mks <- list(Epi=epi_gns, Fib=fib_gns, Peri=peri_gns, Endo=endo_gns, Adipo=adipo_gns,
                Mac=mac_gns, DC=dc_gns, Mast=mast_gns, T=t_gns, NK=nk_gns, B=b_gns, Pl=pl_gns)

mks <- unlist(lst.mks)
names(mks) <- c()
table(mks %in% all_gns)  # Check if all markers are in the dataset
```

## Check cell types
### [Fig S7A] celltype on UMAP
```{r}
png(file.path(fig_dir1, "umap_wp_celltype_wo_other.png"), width = 700,height = 700)
p <- DimPlot(seurat_all, group.by = "celltype_broad", label = FALSE, reduction = "umap", raster=FALSE, pt.size=0.1) +
  theme(legend.text = element_text(size = 24),
        legend.key.size=unit(5,"mm")) +
  labs(title = NULL)
print(p)
dev.off()
```

### [Fig S7B] DotPlot
```{r}
cts <- seurat_all$celltype_broad
seurat_all@meta.data$celltype_broad_rev <- factor(seurat_all@meta.data$celltype_broad,levels=rev(levels(cts)))

pdf(file.path(fig_dir1,"dotplot_wp_celltype.pdf"),width=12,height=6)
DotPlot(seurat_all, features = unique(mks), group.by = "celltype_broad_rev") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)) +
  xlab("") + ylab("")
dev.off()

```

### [Fig S7H] Compute correlation between cell type frequency and module expression (bulk)
```{r}
df.gs <- readRDS(file.path(obj_dir,"DFBCC_txn","dfbcc_bulk_mod_activity.rds")) %>%
  mutate(bot = sub("BOT","",bot)) %>%
  arrange(desc(ERS)) %>%
  filter(bot %in% unique(seurat_all$sample))
bots.used <- df.gs$bot
new.ids <- df.gs$new

seurat_all$sample <- factor(seurat_all$sample, levels=bots.used)

##
set.seed(123)
tab <- table(seurat_all$celltype_broad,seurat_all$sample)[-13,bots.used]
rownames(tab)[6] <- "Mono_Mac"
names(ct.cols) <- rownames(tab)
mat.freq <- array(tab,dim(tab))
nfreq <- apply(mat.freq,2,function(x)x/sum(x))

rownames(nfreq) <- rownames(tab)
colnames(nfreq) <- new.ids # colnames(tab)  

pdf(file.path(fig_dir1,"cts_barplot_dfbcc.pdf"),width=7,height=7)

par(mar=c(5,5,4,10))
barplot(nfreq[12:1,],col=rev(exp.ct.cols2[1:12]),las=2,
        beside=F,border=FALSE,
        xlab="",ylab="Frequency",
        main="Cell type frequency (DF/BCC, snRNA-seq)",cex.names=1)
par(xpd=T)
legend(par()$usr[2],par()$usr[4],
       names(exp.ct.cols2[1:12]),
       fill=exp.ct.cols2[1:12])
par(xpd=F)

dev.off()
```

```{r}
saveRDS(nfreq, file=file.path(obj_dir1, "dfbcc_snseq_ctfreq.rds"))
```

### [Fig S7I] Correlation between cell type abundance
```{r}
ers <- df.gs$ERS
names(ers) <- df.gs$new

nms <- names(df.gs)[2:5]

cors <- apply(nfreq[1:12,],1,function(x)cor(x,df.gs[2:5]))
rownames(cors) <- nms

cor.pvals <- apply(nfreq[1:12,],1,function(x){
  sapply(df.gs[2:5],function(y){
    tmp <- cor.test(x,y)$p.value
  })
})

is.sig <- cor.pvals < 0.05

# boxplot(abs(cors) ~is.sig)
cor.th <- mean(c(min(abs(cors[is.sig])),max(abs(cors[!is.sig]))))

cor1 <- cor(t(nfreq[1:12,]))

pdf(file.path(fig_dir1,"dfbcc_cts_correlation.pdf"),width=7,height=7)
heatmap3(cor1,balanceColor=T,scale="none",
         main="Cell type frequency correlation (DF/BCC, snRNA-seq)")
dev.off()
```

### [Fig S7X] Correlation between cell type abundance
```{r}
nms <- names(df.gs)[2:5]
rng <- range(cors)
for(nm in nms){
  o <- order(cors[nm,])
  this.cors <- cors[nm,o]
  is.sig.1 <- is.sig[nm,o]
  names(this.cors) <- paste0(names(this.cors),c("","*")[is.sig.1+1])
  pdf(file.path(fig_dir1,paste0("barplot_cor_cts_",sub("/","",nm),".pdf")),width=4,height=7)
  par(mar=c(5,7,4,1))
  barplot(this.cors,horiz=T,col=ct.cols[o],las=1,
          main=paste0(nm," (snRNA-seq)"),xlim=rng)
  abline(v=cor.th*c(-1,1),lty=2)
  abline(v=0)
  dev.off()
}
```
