==> src/16_DFBCC_snSeq_analysis.Rmd <==
---
title: "10_scRNAseq_data_analysis"
output: html_document
date: "2023-09-09"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Initial setup

### Set up paths
```{r}
obj_dir1 <- file.path(obj_dir,"DFBCC_sn")
fig_dir1 <- file.path(fig_dir,"16_DFBCC_snSeq_analysis")

dir.create(file.path(fig_dir1,"mean_activity"))  
dir.create(file.path(fig_dir1,"mod_activity"))
dir.create(file.path(fig_dir1,"umap"))
```

### Load libraries
```{r}
suppressPackageStartupMessages({
  library(dplyr)
  library(heatmap3)
  library(ggplot2)
  library(ggrepel)
  library(Seurat)
  library(viridis)
  library(RColorBrewer)
  library(GSVA)
  library(fgsea)
})
source(file.path(src_dir,"utils","h3.R"))
```

### Load data
#### scRNA-seq data
```{r}
wp <- readRDS(file=file.path(obj_dir1,"seurat_all.rds"))
df.gs <- readRDS(file.path(obj_dir,"DFBCC_txn","dfbcc_bulk_mod_activity.rds"))
```

## Load Module genes
```{r}
mods <-readRDS(file.path(obj_dir,"Gene_modules","module_genes_final.rds"))
common.tested.gns <- readRDS(file.path(obj_dir,"TCGA_META","common_tested_gns.rds"))
```

```{r}
mat0 <- GetAssayData(wp)
all.gns <- rownames(mat0) 
genes_data <- rownames(GetAssayData(object = wp, slot = "data"))

mods <- lapply(mods,function(gns){
  gns[gns %in% all.gns] 
}) ## 250 genes - CXCL5, HLA_H, KIR3DL3, CSF2 - are not used

load(file.path(obj_dir,"misc","tcga_heatmap_sidecols.rda")) # mod.side.cols

ss <- mod.side.cols[,"immune"]
ss[ss=="white"] <- "steelblue"
mod.side.cols[,"immune"] <- ss
mod.side.cols <- mod.side.cols[,-2]
colnames(mod.side.cols) <- c("hallmark","module")

mod.uniq.cols <- c("steelblue",brewer.pal(3,"Accent")[c(1,3,2)])#f3756c","#7caf42","#16bcc2","#a780ba")
names(mod.uniq.cols) <- names(mods)
```

## Compute module expression in individual cells with GSVA - `gsva_scores`
```{r}
used.all.gns <- intersect(all.gns,common.tested.gns) # 13423 genes

mat <- mat0[used.all.gns,]

if(0){
  system.time({
    ssgsea.par <- ssgseaParam(expr = mat, geneSets = mods, minSize = 15, maxSize = 500)
    gsva_scores <- gsva(ssgsea.par)
    gsva_scores <- t(gsva_scores)
  })
  saveRDS(gsva_scores,file.path(obj_dir1,"scores_dfci_trimmed_mat.rds"))
}else{
  gsva_scores <- readRDS(file.path(obj_dir1,"scores_dfci_trimmed_mat.rds"))
}

wp@meta.data$ERS <- gsva_scores[,"ERS"]
wp@meta.data$`TNFa/NFkB` <- gsva_scores[,"TNFa/NFkB"]
wp@meta.data$`IFN-I` <- gsva_scores[,"IFN-I"]
wp@meta.data$`APM/TC` <- gsva_scores[,"APM/TC"]
```

### [Fig S7B] Color UMAP with module activities
```{r}
for(mod in names(mods)){
  c1 <- quantile(wp@meta.data[[mod]],0.01)
  c2 <- quantile(wp@meta.data[[mod]],0.99)
  ttl <- sub("[-/]","",mod)
  p <- FeaturePlot(wp, reduction = "umap", features = mod,raster=FALSE,
              min.cutoff=c1,max.cutoff=c2,pt.size=0.1) + 
    scale_color_viridis(option = "D", direction = 1) +
    NoLegend() +
    labs(title=mod)
  png(file.path(fig_dir1,"umap",paste0("umap_wp_",ttl,".png")),width=700,height=700)
  print(p)
  dev.off()
}

```

### Estimate activity scores accurately - linear regression model - per patient & per cell type (`pts`, `lst.mean.mods`, `cts1`)
```{r}
mod.gns <- unlist(mods)
gn.exp <- mat[mod.gns,] %>%
  as.matrix() %>% 
  t() %>%
  as.data.frame()# %>%

## rename Epithelial -> Cancer
names(wp@meta.data)

cts <- wp$celltype_broad
wp$celltype_2 <- cts

uniq.bot <- df.gs$bot
names(uniq.bot) <- paste0("df",seq(uniq.bot))

# plot(match(paste0("BOT",levels(wp$sample)),uniq.bot))
uniq.bot.1 <- uniq.bot[uniq.bot %in% paste0("BOT",unique(wp$sample))]
new.id.1 <- names(uniq.bot.1)
names(new.id.1) <- uniq.bot.1

gn.exp.mods <- gn.exp %>%
  mutate(bot=factor(paste0("BOT",wp$sample),uniq.bot.1)) %>%
  mutate(smpl=new.id.1[as.character(bot)]) %>%
  mutate(smpl=factor(smpl,levels=new.id.1)) %>%
  mutate(celltype=wp$celltype_2) %>%
  mutate(subtype=wp$subtype)

names(gn.exp.mods)<- c(mod.gns,"bot","smpl","celltype","subtype")

## identical(rownames(gn.exp.mods),rownames(scores)) ### TRUE

lst.mean.gns <- lapply(gn.exp.mods[mod.gns],function(sc){
  dfs <- as.data.frame(tapply(sc,list(gn.exp.mods$smpl,gn.exp.mods$celltype),mean,na.rm=T))[1:12]
})

lst.mean.mods <- apply(gsva_scores,2,function(sc){
  dfs <- as.data.frame(tapply(sc,list(gn.exp.mods$smpl,gn.exp.mods$celltype),mean,na.rm=T))[1:12]
},simplify=FALSE)
```

### [Fig S7D] comparison of expression pattern of module genes
```{r}
df.all <- do.call(cbind,lapply(lst.mean.gns,unlist))
cor.all.gns <- cor(df.all,use="pairwise.complete.obs")

h <- heatmap3(cor.all.gns,balanceCol=TRUE,scale="none",
         ColSideColors=mod.cols[rownames(cor.all.gns)],
         RowSideColors=mod.cols[rownames(cor.all.gns)])
h.ri <- h$rowInd

cut10 <- cutree(h$hcr,10)
uniq.cut10 <- unique(cut10[h$rowInd])
cut10 <- as.numeric(factor(cut10,levels=uniq.cut10))
# identical(rownames(rc),names(cut10))
set.seed(12)
rc <- cbind(mod.side.cols[rownames(cor.all.gns),],celltype=sample(brewer.pal(10,"Spectral"))[cut10])

##
pdf(file.path(fig_dir1,"mod_activity","heatmap_250_mod_gene_exp-ptrn.pdf"),width=7,height=7)
h <- heatmap3(cor.all.gns,balanceCol=TRUE,scale="none",
         ColSideColors=rc,RowSideColors=rc)
dev.off()
```


```{r}
mods.v <- rep(names(mods),sapply(mods,length))
names(mods.v) <- unlist(mods)
# identical(names(mods.v),names(cut10))
mod.cl <- table(module=mods.v,cluster=cut10)
```


#### [Fig S7D - cont'd] Compute median level - require `mods`, `mod.side.cols`, `mat0`, `cts1`
```{r}
uniq.gns <- unlist(mods)
rc <- mod.side.cols[uniq.gns,]
mat1 <- mat0[uniq.gns,]

cts1 <- wp$celltype_2
mean.mat <- t(apply(mat1,1,function(x){
  tmp <- tapply(x,cts1,mean,na.rm=T)
  tmp <- tmp/max(tmp,na.rm=T)
  return(tmp)
}))

wm <- apply(mean.mat,1,which.max)

mean.mat.1 <- t(apply(mat1,1,function(x){
  tmp <- tapply(x,cts1,mean,na.rm=T)
  # tmp <- tmp/max(tmp,na.rm=T)
  return(tmp)
}))

used.bot <- sub("BOT","",df.gs$bot)
names(used.bot) <- df.gs$new
smpls <- factor(wp$sample,levels=used.bot[used.bot %in% wp$sample])
mean.mat.s <- t(sapply(uniq.gns,function(gn){
  x <- mat1[gn,]
  # this.ct <- levels(cts1)[wm[gn]]
  # tmp <- tapply(x,list(smpls,cts1),mean,na.rm=T)#[,this.ct]
  tmp <- tapply(x,smpls,mean,na.rm=T)
  tmp <- tmp/max(tmp,na.rm=T)
  return(tmp)
}))

rownames(mean.mat.s) <- uniq.gns

```

#### [Fig S7D mid/right] heatmap relative expression among cell types for each gene

```{r}
cols <- viridis(100)

pdf(file.path(fig_dir1,"mod_activity",paste0("heatmap_ct_expression_all.pdf")),width=7,height=7)
xh <- h3(mean.mat[h.ri,1:12],scale="none",
     RowSideColors=rc[h.ri,],col=cols,Rowv=NA,Colv=NA)
dev.off()

pdf(file.path(fig_dir1,"mod_activity",paste0("heatmap_smpl_expression_all.pdf")),width=7,height=7)
xh <- h3(mean.mat.s[h.ri,],scale="none",
     RowSideColors=rc[h.ri,],col=cols,Rowv=NA,Colv=NA)
dev.off()
```

### [Fig S7D far-right] representing cell-type per cluster - require `cut10`,`rc`

```{r}
tum.exp <- mean.mat[,"Cancer"]
tum.exp.1 <- mean.mat.1[,"Cancer"] ## already log scale

max.tis <- colnames(mean.mat)[apply(mean.mat,1,which.max)]
names(max.tis) <- rownames(mean.mat)

df.tum.exp <- data.frame(genes=names(tum.exp),
                         tum.exp=tum.exp,
                         tum.exp.1=tum.exp.1,
                         celltype=factor(cut10),
                         exp.tis=max.tis,
                         module=mods.v)

n.clusts <- 10
xh <- h3(mean.mat[h.ri,],scale="none",
     RowSideColors=rc,col=cols,Colv=NA,
     distfunR=dist)
hc <- xh$hcr
mem <- cutree(hc,k=n.clusts)
mem <- mem[xh$rowInd]
mem <- factor(mem,levels=rev(unique(mem)))

lst.gns.by.exp.cts <- tapply(names(mem),mem,identity)
names(lst.gns.by.exp.cts) <- seq(lst.gns.by.exp.cts)

lst.gns.by.cor <- tapply(df.tum.exp$genes,df.tum.exp$celltype,identity)

th1 <- 0.7 # th to consider if each celltype expresses`each gene` 
th2 <- 0.3 # th to consider if each celltype expresses `this cluster of genes`

ct.exp <- lapply(lst.gns.by.cor,function(x){
  tmp <- mean.mat[x,]
  oc <- order(colMeans(tmp),decreasing=T)
  tmp <- tmp[,oc]
  nm <- colnames(tmp)
  n <- length(x)
  is.exp <- factor(unlist(apply(tmp,1,function(y){
    names(which(y > th1))
  })),levels=nm)
  is.exp <- table(is.exp)
  
  this.cts <- names(which(sort(is.exp,decreasing=T)/n > th2))
  # this.cts <- sort(is.exp,decreasing=T)/n
  return(this.cts)
})

print(ct.exp)
```

### [Fig S7F] Compute the similarity of the expression pattern of each gene
```{r}
df.each <- do.call(cbind,lapply(lst.mean.mods,as.matrix))
colnames(df.each) <- paste(rep(names(lst.mean.mods),sapply(lst.mean.mods,ncol)),colnames(df.each),sep="_")
cor.each <- cor(df.each,use="pairwise.complete.obs",method="pearson")

mod.cols <- rep(mod.uniq.cols,sapply(lst.mean.mods,ncol))

pdf(file.path(fig_dir1,"mod_activity","heatmap_mod_activity.pdf"),width=7,height=7)
h3(cor.each,balanceCol=TRUE,scale="none",
   ColSideColors=cbind(module=mod.cols),
   RowSideColors=cbind(module=mod.cols),
   margins=c(10,10))
dev.off()

```

## Module expression per cell type and per sample - `coef.cts.1` and `coef.pts.1`
```{r}
coef.cts <- coef.pts <- list()
nms.mods <- names(mods)

used.cts <- names(lst.mean.mods$ERS)

for(nm in nms.mods){
  mat <- lst.mean.mods[[nm]]

  if(nm == "APM/TC"){
    mat1 <- mat[,!names(mat) %in% c("Cancer","Endo","Fibro","Peri")]
  }else{
    mat1 <- mat
  }

  tmp1 <- data.frame(log_activity = log(unlist(mat1)), # activity = unlist(mat1), 
                    pt = factor(rep(rownames(mat1),times=ncol(mat1))),
                    ct = factor(rep(colnames(mat1),each=nrow(mat1))),levels=names(mat1)) %>%
    filter(!is.na(log_activity)) # filter(!is.na(activity)) 
  lm1 <- lm(log_activity ~ pt + ct + 0, data=tmp1)
  coef <- summary(lm1)$coefficients[,1]
  
  coef.pt <- coef[grep("^pt",names(coef))]
  sorted.pts <- sub("pt","",names(sort(coef.pt)))
  coef.pts[[nm]] <- coef.pt
  
  coef.ct <- c(ctAdipo=0,coef[grep("^ct",names(coef))])

  if(nm=="APM/TC"){
    names(coef.pt) <- sub("pt","",names(coef.pt))
    tmp <- data.frame(log_activity = log(unlist(mat)), # activity = unlist(mat), 
                  pt = factor(rep(rownames(mat),times=ncol(mat))),
                  ct = factor(rep(colnames(mat),each=nrow(mat))),levels=names(mat)) %>%
           filter(!is.na(log_activity)) %>%
           mutate(log_activity_1 = log_activity + coef.pt[pt])
    lm2 <- lm(log_activity_1 ~ ct + 0, data=tmp)
    coef <- summary(lm2)$coefficients[,1]
    coef <- coef - coef["ctB"]
    coef.ct <- c(coef[grep("^ct",names(coef))])

  }
  names(coef.ct) <- sub("ct","",names(coef.ct))
  coef.ct <- coef.ct[names(mat)]

  this.ct <- rep(NA,length(used.cts))
  names(this.ct) <- used.cts
  this.ct[names(coef.ct)] <- coef.ct
  
  coef.cts[[nm]] <- this.ct
}

coef.cts.1 <- as.data.frame(exp(do.call(cbind,coef.cts)))
coef.pts.1 <- as.data.frame(exp(do.call(cbind,coef.pts)))

## reorder coef.pts.1
rownames(coef.pts.1) <- sub("^pt","",rownames(coef.pts.1))
o <- order(as.numeric(sub("df","",rownames(coef.pts.1))))
df.pts <- coef.pts.1[o,] %>% tibble::rownames_to_column("new")
  
save(df.pts,file=file.path(obj_dir1,"mod_activity_dfbcc_all.rda"))
```

### [Fig S7E] Heatmap - compute avg score per `ct` and `pt` - sanity check - require `lst.mean.mods`, `cts1`
```{r}
cts <- used.cts
cts1 <- factor(cts,levels=cts)

sorted.pts <- df.pts$new

for(nm in nms.mods){
  mat <- lst.mean.mods[[nm]]
  old.pts <- rownames(mat)

  # rownames(mat) <- df.pts$new[match(rownames(mat),df.pts$old)] 
  cols <- viridis(100)
  pdf(file.path(fig_dir1,"mod_activity",paste0("avgexp_",sub("/","",nm),"_raw.pdf")),
      width=7,height=7)
  h3(mat,balanceColor=FALSE,Colv=NA,Rowv=NA,
     scale="none",na.color="grey",col=viridis(100),cexRow=.8)

  mtext(paste0("Activity of ",nm," module\n(No scaling)"),
    side=3,line=0,cex=2)
  dev.off()
}
```

#### [Fig S7G] heatmap - patient and cell type level activities
```{r}
# cell type level
pdf(file.path(fig_dir1,"mod_activity","coef_cts_raw.pdf"),width=7,height=7)
h3(coef.cts.1[rev(seq(nrow(coef.cts.1))),],balanceColor=FALSE,Colv=NA,Rowv=NA,scale="column",
   na.color="grey",col=cols,
   margins=c(10,5))
dev.off()

## patient-level
mat.pts <- as.matrix(df.pts[-1])
rownames(mat.pts) <- as.character(df.pts$new)

pdf(file.path(fig_dir1,"mod_activity","coef_pts_raw.pdf"),width=7,height=7)
h3(mat.pts,balanceColor=FALSE,Colv=NA,Rowv=NA,scale="column",na.color="grey",
   col=cols,margins=c(10,5),cexRow = .8)
dev.off()
```

#### [Fig S7K] correlation among all subtypes or within HR+
```{r}
cors.all <- cor(coef.pts.1,method="pearson",use="pairwise.complete.obs")
pdf(file.path(fig_dir1,"mod_activity","correlation_all_pts.pdf"),width=7,height=7)
h3(cors.all,balanceCol=TRUE,scale="none",Colv=NA,Rowv=NA,na.col="grey",
   ColSideColors=mod.uniq.cols,RowSideColors=mod.uniq.cols,
   main="PCC between module activity (per patient), all subtypes")
dev.off()

write.csv(round(cors.all,3),file.path(fig_dir1,"mod_activity","correlation_all_pts.csv"))
```


## [Fig S7J] correlation between activity and cell types

```{r}
nfreq <- readRDS(file.path(obj_dir1,"dfbcc_snseq_ctfreq.rds"))
cors.all <- cor(t(nfreq[,df.pts$new]),df.pts[names(mods)],use="pairwise.complete.obs",method="pearson")

for(nm in nms){
  nm1 <- sub("[-/]","",nm)
  cors.all.mod <- sort(cors.all[,nm],decreasing=F)
  pdf(file.path(fig_dir1,"mod_activity",paste0("barplot_cor_all_",nm1,"_102524.pdf")),width=7,height=7)
  par(mar=c(5,10,4,1))
  barplot(cors.all.mod,las=1,horiz=TRUE,col=exp.ct.cols2[names(cors.all.mod)],
          main=paste0("PCC between ",nm," activity and cell type frequency"),
          xlab="PCC",ylab="")
  dev.off()
}

```

## [Fig S7C] linear regression
```{r}
coef.pts.2 <- df.pts
names(coef.pts.2)[2:5] <- paste0("sn_",names(df.pts)[2:5])

names(df.gs)[2:5] <- paste0("bulk_",names(df.gs)[2:5])

df.ers <- inner_join(coef.pts.2,df.gs,by="new")

lm1 <- lm(bulk_ERS ~ sn_ERS, data=df.ers)

pdf(file.path(fig_dir1,"mod_activity","plot_ERS_activity_comparison.pdf"),width=7,height=7)
par(mar=c(5,5,3,3))
cor.1 <- cor(df.ers$sn_ERS,df.ers$bulk_ERS,use="pairwise.complete.obs",method="pearson")
plot(df.ers$sn_ERS,df.ers$bulk_ERS,
     xlab="Mean ERS activity in tumor cells (sn)",ylab="Bulk ERS activity",
     main=paste0("ERS activity in two platforms", " (PCC=",round(cor.1,2),")"),
     pch=20)
abline(lm1,col="red")
dev.off()
```
