==> src/20_DFBCC_combo.Rmd <==
---
title: "10_scRNAseq_data_analysis"
#output: html_document
date: "2023-09-09"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Initial setup

### Load libraries
```{r}
suppressPackageStartupMessages({
  library(dplyr)
  library(Matrix)
  library(data.table)
  library(ggplot2)
  library(ggrepel)
  library(Seurat)
  library(RColorBrewer)
  library(fgsea)
  library(GSVA)
})
```

### Load module genes
```{r}
## whole tumor
mods <- readRDS(file.path(obj_dir,"Gene_modules","module_genes_final.rds"))
names(mods) <- paste0(names(mods),"_whole")

## cancer cell intrinsic
cons.gns <- readRDS(file.path(obj_dir,"CCLE","conserved_gns.rds"))
names(cons.gns) <- paste0(names(cons.gns),"_int")

## intrinsic sub
nfkb_sub_1 <- intersect(cons.gns$`TNFa/NFkB_int`,mods$`TNFa/NFkB_whole`)
nfkb_sub_2 <- intersect(cons.gns$`TNFa/NFkB_int`,mods$`APM/TC_whole`)
ifn_sub <- intersect(cons.gns$`IFN-I_int`,mods$`IFN-I_whole`)
apm_sub <- intersect(cons.gns$`IFN-I_int`,mods$`APM/TC_whole`)

msigdb <- readRDS(file.path(obj_dir,"Pathways","msigdb.rds"))

mods.1 <- c(mods,cons.gns[1:3],
              list(NFkB_sub1 = nfkb_sub_1,
                   NFkB_sub2 = nfkb_sub_2,
                   IFN_sub = ifn_sub,
                   APM_sub = apm_sub))
```

## Correlation between cancer cell frequency and transcriptome
## Load txn data 
```{r}
wehi.sc <- readRDS(file.path(obj_dir,"WEHI","wehi_seurat_all_cells.rds"))
dfbcc.sn <- readRDS(file=file.path(obj_dir,"DFBCC_sn","seurat_all.rds"))
dfbcc.bulk <- readRDS(file.path(obj_dir,"DFBCC_txn","dfbcc_bulk_logtxn.rds"))
```

#### common genes
```{r}
wehi.sc1 <- GetAssayData(wehi.sc)
gns1 <- rownames(wehi.sc1)

dfbcc.sn1 <- GetAssayData(dfbcc.sn)
gns2 <- rownames(dfbcc.sn1)
gns3 <- rownames(dfbcc.bulk)

all.syms <- unique(unlist(msigdb))

# txn.gns <- intersect(intersect(gns1,gns2),gns3) # 14954 genes
txn.gns <- intersect(gns1,gns2) # 22171 genes 

common.used.syms <- intersect(txn.gns,all.syms) # 15573 genes

if(0){
  saveRDS(common.used.syms,file=file.path(obj_dir,"DFBCC_combo","common.used.syms.rds"))
}else{
  common.used.syms <- readRDS(file=file.path(obj_dir,"DFBCC_combo","common.used.syms.rds"))
}
```

## Filter out genes not expressed in cancer cells (WEHI, DFBCC)

### WEHI (all samples)
```{r}
mat0 <- GetAssayData(wehi.sc)
mat1 <- mat0[common.used.syms,]

m.cts <- list()
uniq.cts <- levels(wehi.sc$celltype_2)
for(ct in uniq.cts){
  cat(ct,"..\n")
  is.ct <- wehi.sc$celltype_2 == ct
  ct.smpls <- wehi.sc$orig.ident[is.ct]
# Operate on the sparse matrix directly
  mat2 <- mat1[, is.ct]  # Keep the matrix in sparse format (do not convert)
  
  # Compute the row-wise mean grouped by ct.smpls without converting to dense matrix
  m.ct <- list()
  for(smpl in unique(ct.smpls)){
    cat(smpl," ")
    tmp <- rowMeans(mat2[, ct.smpls == smpl, drop=FALSE])  # Compute row means on sparse matrix
    m.ct[[smpl]] <- tmp
  }
  m.ct <- do.call(cbind, m.ct)
  
  if(ct == "Cancer"){
    m.tum <- m.ct
    saveRDS(m.tum,file=file.path(obj_dir,"DFBCC_combo","m.tum_wehi.rds"))
  }
  # Store the result for the current cell type
  m.cts[[ct]] <- m.ct
  cat("\n")
}

exp.gs <- sapply(m.cts,function(m){
  rowMeans(m,na.rm=T)
})

ratio.tum <- apply(exp.gs,1,function(e){
  e[1]/max(e,na.rm=T)
})

is.exp.by.tum <- ratio.tum > .1 # 13411 TRUE, 2162 FALSE
saveRDS(is.exp.by.tum,file=file.path(obj_dir,"DFBCC_combo","is_exp_by_tum_wehi.rds"))
```

### DFBCC data
```{r}
mat0 <- GetAssayData(dfbcc.sn)
mat1 <- mat0[common.used.syms,]

m.cts <- list()
uniq.cts <- levels(dfbcc.sn$celltype_broad)
for(ct in uniq.cts){
  cat(ct,"..\n")
  is.ct <- dfbcc.sn$celltype_broad == ct
  ct.smpls <- dfbcc.sn$sample[is.ct]
# Operate on the sparse matrix directly
  mat2 <- mat1[, is.ct]  # Keep the matrix in sparse format (do not convert)
  
  # Compute the row-wise mean grouped by ct.smpls without converting to dense matrix
  m.ct <- list()
  for(smpl in unique(ct.smpls)){
    cat(smpl," ")
    tmp <- rowMeans(mat2[, ct.smpls == smpl, drop=FALSE])  # Compute row means on sparse matrix
    m.ct[[smpl]] <- tmp
  }
  m.ct <- do.call(cbind, m.ct)
  if(ct == "Cancer"){
    m.tum <- m.ct
    saveRDS(m.tum,file=file.path(obj_dir,"DFBCC_combo","m.tum_dfbcc.rds"))
  }
  
  # Store the result for the current cell type
  m.cts[[ct]] <- m.ct
}

exp.gs <- sapply(m.cts,function(m){
  rowMeans(m,na.rm=T)
})

ratio.tum <- apply(exp.gs,1,function(e){
  e[1]/max(e,na.rm=T)
})

is.exp.by.tum <- ratio.tum > .1 # 12940 TRUE, 2633 FALSE
saveRDS(is.exp.by.tum,file=file.path(obj_dir,"DFBCC_combo","is_exp_by_tum_dfbcc.rds"))
```

### assemble genes expressed in cancer cells
```{r}
common.used.syms <- readRDS(file=file.path(obj_dir,"DFBCC_combo","common.used.syms.rds"))
is.exp.wehi <- readRDS(file=file.path(obj_dir,"DFBCC_combo","is_exp_by_tum_wehi.rds"))
is.exp.dfbcc <- readRDS(file=file.path(obj_dir,"DFBCC_combo","is_exp_by_tum_dfbcc.rds"))

is.exp.by.tum <- is.exp.wehi & is.exp.dfbcc
common.syms <- common.used.syms[is.exp.by.tum] ## 14033 genes
common.syms <- common.syms[!is.na(common.syms)]

m.tum.wehi <- readRDS(file=file.path(obj_dir,"DFBCC_combo","m.tum_wehi.rds"))[common.syms,]
m.tum.dfbcc <- readRDS(file=file.path(obj_dir,"DFBCC_combo","m.tum_dfbcc.rds"))[common.syms,]
```

## Correlation between gene expression and cancer cell frequency
### WEHI (all samples, HR+ only)
```{r}
wehi.sc$subtype[is.na(wehi.sc$subtype)] <- "HR+"
subtypes <- wehi.sc@meta.data[c("subtype","orig.ident")] %>% unique()
is.hr <- subtypes %>% filter(subtype == "HR+") %>% pull(orig.ident) %>% as.character()

## cell type frequency
tab <- as.matrix(table(wehi.sc$celltype_2,wehi.sc$orig.ident))
stat <- array(as.vector(tab),dim(tab),dimnames=dimnames(tab))
nstat <- stat/colSums(stat)[col(stat)] 
ccf.wehi <- nstat["Cancer",]

cor.wehi.all <- cor(t(m.tum.wehi),ccf.wehi,method="pe")[,1]
cor.wehi.hr <- cor(t(m.tum.wehi[,is.hr]),ccf.wehi[is.hr],method="pe")[,1]

saveRDS(cor.wehi.all,file=file.path(obj_dir,"DFBCC_combo","cor_wehi_all.rds"))
saveRDS(cor.wehi.hr,file=file.path(obj_dir,"DFBCC_combo","cor_wehi_hr.rds"))
```
### DFBCC (snRNA-seq)
```{r}
tab <- as.matrix(table(dfbcc.sn$celltype_broad,dfbcc.sn$sample))
stat <- array(as.vector(tab),dim(tab),dimnames=dimnames(tab))
nstat <- stat/colSums(stat)[col(stat)] 

ccf.dfbcc.sn <- nstat["Cancer",]
cor.dfbcc.sn <- cor(t(m.tum.dfbcc),ccf.dfbcc.sn,method="pe")[,1]
saveRDS(cor.dfbcc.sn,file=file.path(obj_dir,"DFBCC_combo","cor_dfbcc.sn.rds"))
```

### DFBCC (CyCIF)
```{r}
nsh <- readRDS(file=file.path(obj_dir,"DFBCC_cycif","nstat_cycif.rds"))
df.gs <- readRDS(file.path(obj_dir,"DFBCC_txn","dfbcc_bulk_mod_activity.rds"))
bots <- sub("BOT","",df.gs$bot)
names(bots) <- df.gs$new
colnames(nsh) <- bots[colnames(nsh)]
ccf.dfbcc.cy <- nsh["Cancer",]

smpls <- colnames(m.tum.dfbcc)
cor.dfbcc.cy <- cor(t(m.tum.dfbcc[,smpls]),ccf.dfbcc.cy[smpls],method="pe")[,1]
saveRDS(cor.dfbcc.cy,file=file.path(obj_dir,"DFBCC_combo","cor_dfbcc.cy.rds"))
```

### DFBCC (bulk vs CyCIF)
```{r}
colnames(dfbcc.bulk) <- sub("BOT","",names(dfbcc.bulk))
smpls <- colnames(m.tum.dfbcc)

cor.dfbcc.bulk.cy <- cor(t(dfbcc.bulk[common.syms,smpls]),ccf.dfbcc.cy[smpls],method="pe")[,1]
saveRDS(cor.dfbcc.bulk.cy,file=file.path(obj_dir,"DFBCC_combo","cor_dfbcc.bulk.cy.rds"))
```

## Summary - correlations
```{r}
cor.wehi <- readRDS(file=file.path(obj_dir,"DFBCC_combo","cor_wehi_all.rds"))
cor.dfbcc <- readRDS(file=file.path(obj_dir,"DFBCC_combo","cor_dfbcc.cy.rds"))
cor.dfbcc.bulk <- readRDS(file=file.path(obj_dir,"DFBCC_combo","cor_dfbcc.bulk.cy.rds"))
cors <- list(
  wehi = cor.wehi,
  dfbcc = cor.dfbcc,
  dfbcc.bulk = cor.dfbcc.bulk
)
```

```{r}
save(cors,file=file.path(obj_dir,"DFBCC_combo","cors_wehi_dfci.rda"))
```

#### Pathway analysis - fgsea
```{r}
msigdb0 <- lapply(msigdb,function(x)intersect(x,common.syms))
n <- sapply(msigdb0,length) 
is.used <- n > 15 & n < 500# 5298
msigdb.used <- msigdb0[is.used]

msigdb0.1 <- lapply(mods.1,function(x)intersect(x,common.syms))
n <- sapply(msigdb0.1,length) 
is.used <- n > 3 & n < 500# 5298
msigdb.used.1 <- msigdb0.1[is.used]

fg <- list()
for(nm in names(cors)){
  cors1 <- sort(cors[[nm]],decreasing=F)
  fg1 <- fgsea(pathways = msigdb.used,
                    stats    = cors1,
                    minSize  = 15,
                    maxSize  = 500) %>%
    mutate(slp = -log10(padj) * sign(ES))
  fg1.1 <- fgsea(pathways = msigdb.used.1, 
                    stats    = cors1,
                    minSize  = 3,
                    maxSize  = 500) %>%
    mutate(slp = -log10(padj) * sign(ES))
  fg[[nm]] <- rbind(fg1,fg1.1)
}

idx <- lapply(fg,function(x)x[[1]])
is.tested <- idx[[1]] %in% idx[[2]]

fg1 <- lapply(fg,function(x){
  is.tested <- x[[1]] %in% fg[[2]][[1]]
  return(x[is.tested,])
})
```

### [Fig 4F] plot fgsea results - cancer cell specific
```{r}
df.fg <- data.frame(
  `WEHI` = fg1$wehi$slp,
  `DF/BCC` = fg1$dfbcc$slp,
  pathway = fg1$wehi$pathway
) %>%
  mutate(type = rep(c(0,1,3,2,1,2,1,3,2,1),c(5514,1,1,1,1,1,1,1,1,1)))

my.uniq.cols <- brewer.pal(3,"Set1")[c(2,1,3)]
names(my.uniq.cols) <- c("whole tumor","intrinsic","sub_int")

my.cols <- my.uniq.cols[c(1,3,2,1,2,1,3,2,1)]

##
pdf(file.path(fig_dir,"20_DFBCC_combo",paste0("fgsea_sc_vs_sn_tum-gns.pdf")),width=7,height=7)
par(mar=c(5,5,5,5))
plot(df.fg[1:2],pch='.',type="n",asp=1,
     main="Pathway associated with cancer cell frequency in two cohorts",
     xlab="Pathway expression in cancer cells (WEHI)",
     ylab="Pathway expression in cancer cells (DFCI)",     
     xlim=c(-10,10),ylim=c(-10,10))
abline(h=0,v=0)
points(df.fg %>% filter(type==0) %>% select(1:2),pch=".",col="grey30")
points(df.fg %>% filter(type!=0) %>% select(1:2),pch=20,col=my.cols)
text(df.fg %>% filter(type != 0) %>% select(1:2),df.fg %>% filter(type != 0) %>% pull(pathway),col=1)
legend("topright",legend=c("whole tumor","intrinsic","sub int"),pch=20,col=brewer.pal(3,"Set1")[c(2,1,3)],title="module type")
par(xpd=T)
text(par()$usr[2]+2,par()$usr[4],"More expressed\nin samples\nwith high\ncancer cell freq.",cex=.7,col=2)
text(par()$usr[1]-2,par()$usr[3]-2,"More expressed\nin samples\nwith low\ncancer cell freq.",cex=.7,col=4)
par(xpd=F)
dev.off()
```
### [Fig S8D] plot fgsea results - bulk RNA-seq

```{r}
fg.bulk <- fg1$dfbcc.bulk %>% mutate(type="bulk")
fg.bulk.my <- fg.bulk %>% filter(rep(c(FALSE,TRUE),c(5380,9))) %>%
  mutate(gene.kind=c("whole tumor","intrinsic","sub_int")[c(1,3,2,1,2,1,3,2,1)])

p <- ggplot(fg.bulk,aes(x=type,y=slp)) +
  geom_violin() +
  geom_point(data=fg.bulk.my,aes(y=slp,col=gene.kind),pch=95,size=10) +
  scale_color_manual(values = my.uniq.cols) +
  geom_text_repel(data=fg.bulk.my,aes(label=pathway),size=3) +
  xlab("") +
  theme_minimal() +
  ggtitle("Pathway analysis\n(5304 pathways)")

pdf(file.path(fig_dir,"20_DFBCC_combo",paste0("fgsea_violin_bulk.pdf")),width=3,height=7)
print(p)
dev.off()
```

## ssGSEA

### ratio of non-cancer cells over cancer cells within tumor core
```{r}
load(file=file.path(obj_dir,"DFBCC_cycif","freq_ts.rda"))
ratio.st.tum <- freq.ts[3,]/freq.ts[1,]
names(ratio.st.tum) <- bots[names(ratio.st.tum)]
```

### module expression and ssGSEA
```{r}
mat0 <- GetAssayData(dfbcc.sn)
mat1 <- mat0[common.used.syms,]

system.time({
  ssgsea.par <- ssgseaParam(expr = mat1, geneSets = mods.1, minSize = 3, maxSize = 500)
  mod.scores <- gsva(ssgsea.par)
  mod.scores <- t(mod.scores)
})

is.ct <- dfbcc.sn$celltype_broad == "Cancer"
ct.smpls <- dfbcc.sn$sample[is.ct]
mean.mod.scores <- apply(mod_scores[is.ct,],2,function(e){
  tapply(e,ct.smpls,mean)
})
uniq.smpls <- intersect(names(ratio.st.tum),rownames(mean.mod.scores))

plot(df.gs1[])

cors.cc1 <- cor(mean.mod.scores[uniq.smpls,],freq.tum1[uniq.smpls])
cors.cc <- cor(mean.mod.scores[uniq.smpls,],ratio.st.tum[uniq.smpls])

barplot(t(cors.cc),las=2,main="ratio ST of tum")
barplot(t(cors.cc1),las=2,main="cancer cell freq")

df <- data.frame(ratio.st=cors.cc,freq.tum=cors.cc1)
plot(df)
abline(h=0,v=0)

m <- "IFN-I_int"
plot(mean.mod.scores[uniq.smpls,m],ratio.st.tum[uniq.smpls])
cor(mean.mod.scores[uniq.smpls,m],ratio.st.tum[uniq.smpls],method="sp")
cor(mean.mod.scores[uniq.smpls,m],ratio.st.tum[uniq.smpls],method="pe")

lm1 <- lm(ratio.st.tum[uniq.smpls] ~ mean.mod.scores[uniq.smpls,m])
abline(lm1,col=2)
```

### IFN vs cancer cell frequency
```{r}
df.gs <- cbind(df.gs,bot1 = sub("BOT","",df.gs$bot))
df.gs1 <- df.gs[match(uniq.smpls,df.gs$bot1),] %>% mutate(bot1=factor(bot1,levels=uniq.smpls)) %>% arrange(bot1)
cors.wt <- cor(df.gs1[2:5],ratio.st.tum[uniq.smpls])

plot(mean.mod.scores[uniq.smpls,1],freq.tum1[uniq.smpls])
plot(mean.mod.scores[uniq.smpls,1],df.gs1[[2]])

cor(mean.mod.scores[uniq.smpls,3],df.gs1[[4]],method="pe")
```
